requirements.txt
PyYAML>=6.0
numpy>=1.24
pandas>=2.0
torch>=2.1
torchvision>=0.16
timm>=0.9
PyWavelets>=1.5
einops>=0.7
tqdm>=4.66
tensorboard>=2.15
